renamed columValidator file
