+++ functions.ts
@@ -1,3 +1,3 @@
 //TODO: make the type more user friendly, currently it requires Uint8Array
 export { validateRows } from "./src/validators/row-validator/rowValidator";
-export { validateColumns } from "./src/validators/column-validator/columnValidator.class";
+export { validateColumns } from "./src/validators/column-validator/columnValidator";
