+++ src/validators/column-validator/columnValidator.test.ts
@@ -1,4 +1,4 @@
-import { validateColumns } from "../../../src/validators/column-validator/columnValidator.class";
+import { validateColumns } from "./columnValidator";
 import {
   VALID_DEFAULT_SUDOKU_STRING_2D_4x4,
   VALID_DEFAULT_SUDOKU_STRING_2D_9x9,
